remove streaming output backpressure test
